new errors.rs file
