+++ cedar-policy-core/src/validator/entity_manifest.rs
@@ -31,16 +31,24 @@ use serde_with::serde_as;
 use smol_str::SmolStr;
 use thiserror::Error;
 
-pub(crate) mod manifest_helpers;
 pub(crate) mod analysis;
 #[cfg(test)]
 mod entity_manifest_tests;
 #[cfg(test)]
 mod entity_slice_tests;
+mod errors;
 mod human_format;
 mod loader;
+pub(crate) mod manifest_helpers;
 pub mod slicing;
 mod type_annotations;
+
+// Import errors directly
+pub use crate::validator::entity_manifest::errors::{
+    AccessPathNotFoundError, EntityManifestError, EntityManifestFromJsonError,
+    MismatchedEntityManifestError, MismatchedMissingEntityError, MismatchedNotStrictSchemaError,
+    PartialExpressionError, PartialRequestError, UnsupportedCedarFeatureError,
+};
 
 use crate::validator::entity_manifest::analysis::{
     EntityManifestAnalysisResult, WrappedAccessPaths,
@@ -263,119 +271,6 @@ impl Display for EntityRoot {
             EntityRoot::Var(v) => write!(f, "{v}"),
         }
     }
-}
-
-/// Error when expressions are partial during entity
-/// manifest computation
-// CAUTION: this type is publicly exported in `cedar-policy`.
-// Don't make fields `pub`, don't make breaking changes, and use caution
-// when adding public methods.
-#[derive(Debug, Clone, Error, Hash, Eq, PartialEq)]
-#[error("entity slicing requires fully concrete policies. Got a policy with an unknown expression")]
-pub struct PartialExpressionError {}
-
-impl Diagnostic for PartialExpressionError {}
-
-/// Error when the request is partial during entity
-/// manifest computation
-// CAUTION: this type is publicly exported in `cedar-policy`.
-// Don't make fields `pub`, don't make breaking changes, and use caution
-// when adding public methods.
-#[derive(Debug, Clone, Error, Hash, Eq, PartialEq)]
-#[error("entity slicing requires a fully concrete request. Got a partial request")]
-pub struct PartialRequestError {}
-impl Diagnostic for PartialRequestError {}
-
-/// An error generated by entity slicing.
-#[derive(Debug, Error)]
-pub enum EntityManifestError {
-    /// A validation error was encountered
-    // TODO (#1158) impl Error for ValidationResult (it already is implemented for api::ValidationResult)
-    #[error("a validation error occurred")]
-    Validation(ValidationResult),
-    /// A entities error was encountered
-    #[error(transparent)]
-    Entities(#[from] EntitiesError),
-
-    /// The request was partial
-    #[error(transparent)]
-    PartialRequest(#[from] PartialRequestError),
-    /// A policy was partial
-    #[error(transparent)]
-    PartialExpression(#[from] PartialExpressionError),
-    /// Unsupported feature
-    #[error(transparent)]
-    UnsupportedCedarFeature(#[from] UnsupportedCedarFeatureError),
-}
-
-/// Error when entity manifest analysis cannot handle a Cedar feature
-// CAUTION: this type is publicly exported in `cedar-policy`.
-// Don't make fields `pub`, don't make breaking changes, and use caution
-// when adding public methods.
-#[derive(Debug, Clone, Error, Diagnostic)]
-#[error("entity manifest analysis currently doesn't support Cedar feature: {feature}")]
-pub struct UnsupportedCedarFeatureError {
-    pub(crate) feature: SmolStr,
-}
-
-/// Error when the manifest has an entity the schema lacks.
-// CAUTION: this type is publicly exported in `cedar-policy`.
-// Don't make fields `pub`, don't make breaking changes, and use caution
-// when adding public methods.
-#[derive(Debug, Clone, Error, Hash, Eq, PartialEq)]
-#[error("entity manifest doesn't match schema. Schema is missing entity {entity}. Either you wrote an entity manifest by hand (not recommended) or you are using an out-of-date entity manifest with respect to the schema")]
-pub struct MismatchedMissingEntityError {
-    pub(crate) entity: EntityUID,
-}
-
-/// Error when the schema isn't valid in strict mode.
-// CAUTION: this type is publicly exported in `cedar-policy`.
-// Don't make fields `pub`, don't make breaking changes, and use caution
-// when adding public methods.
-#[derive(Debug, Clone, Error, Hash, Eq, PartialEq)]
-#[error("entity manifests are only compatible with schemas that validate in strict mode. Tried to use an invalid schema with an entity manifest")]
-pub struct MismatchedNotStrictSchemaError {}
-
-/// An error generated by entity manifest parsing. These happen
-/// when the entity manifest doesn't conform to the schema.
-/// Either the user wrote an entity manifest by hand (not reccomended)
-/// or they used an out-of-date entity manifest (after updating the schema).
-/// Warning: This error is not guaranteed to happen, even when an entity
-/// manifest is out-of-date with respect to a schema! Users must ensure
-/// that entity manifests are in-sync with the schema and policies.
-// CAUTION: this type is publicly exported in `cedar-policy`.
-// Don't make fields `pub`, don't make breaking changes, and use caution
-// when adding public methods.
-#[derive(Debug, Clone, Error, Hash, Eq, PartialEq)]
-pub enum MismatchedEntityManifestError {
-    /// Mismatch between entity in manifest and schema
-    #[error(transparent)]
-    MismatchedMissingEntity(#[from] MismatchedMissingEntityError),
-    /// Found a schema that isn't valid in strict mode
-    #[error(transparent)]
-    MismatchedNotStrictSchema(#[from] MismatchedNotStrictSchemaError),
-    /// Used wrong entity manifest
-    #[error(transparent)]
-    AccessPathNotFoundError(#[from] AccessPathNotFoundError),
-}
-
-/// An error generated when parsing entity manifests from json
-#[derive(Debug, Error)]
-pub enum EntityManifestFromJsonError {
-    /// A Serde error happened
-    #[error(transparent)]
-    SerdeJsonParseError(#[from] serde_json::Error),
-    /// A mismatched entity manifest error
-    #[error(transparent)]
-    MismatchedEntityManifest(#[from] MismatchedEntityManifestError),
-}
-
-/// Error when an access path is not found in the entity manifest
-#[derive(Debug, Clone, Error, Hash, Eq, PartialEq)]
-#[error("access path not found in entity manifest. This may indicate that you are using the wrong entity manifest with this path")]
-pub struct AccessPathNotFoundError {
-    /// The ID of the path that was not found
-    pub(crate) path_id: usize,
 }
 
 impl AccessPath {
